Match StarRocks integer arithmetic and wrap integer SUM
Integer + - * widen the common operand type one level (TINYINT to SMALLINT
to INT to BIGINT) and wrap at BIGINT; unary minus and ABS widen the same
way. DIV and % keep the common type, never trap on MIN DIV -1, and return
NULL for a zero divisor. Integer literals in arithmetic take their
narrowest type, so SELECT 2147483647 + 1 is BIGINT 2147483648.

SUM over integers up to BIGINT now returns BIGINT and wraps. Every path
accumulates exactly in i128 and truncates at emit, which equals the
wrapped sum, and ORDER BY/HAVING rank the emitted value. The affine SUM
reduction no longer needs an overflow guard.

Because int_col + 1 is now BIGINT, a scalar function whose parameter is
INT (date_add, substr, INTERVAL n MONTH) would stop resolving. When no
overload is reachable by widening, the resolver now narrows an integer
argument, saturating, as StarRocks casts function arguments.

This replaces the PR #79 rule that raised ArithmeticOverflow for integer
arithmetic. Decimal arithmetic keeps raising it.

+++ tests/integration/case_when_test.zig
@@ -231,9 +231,9 @@ test "case: CASE expression continued by arithmetic over a window call" {
     defer q.deinit();
     const batch = (try q.next()).?;
     try std.testing.expectEqual(@as(usize, 3), batch.row_count);
-    try std.testing.expectEqual(@as(i32, 0), batch.values[1].data.int[0]);
-    try std.testing.expectEqual(@as(i32, 1), batch.values[1].data.int[1]);
-    try std.testing.expectEqual(@as(i32, 0), batch.values[1].data.int[2]);
+    try std.testing.expectEqual(@as(i64, 0), batch.values[1].data.bigint[0]);
+    try std.testing.expectEqual(@as(i64, 1), batch.values[1].data.bigint[1]);
+    try std.testing.expectEqual(@as(i64, 0), batch.values[1].data.bigint[2]);
 }
 
 test "case: bare NULL branch adopts the typed branches' type" {
